move the Model contract completion to an abstract parent of our Models
