+++ src/Models/KeyValue.php
@@ -1,13 +1,13 @@
 <?php
+
 /**
- * @copyright 2019 Wayfair LLC - All rights reserved
+ * @copyright 2020 Wayfair LLC - All rights reserved
  */
 
 namespace Wayfair\Models;
 
-use Plenty\Modules\Plugin\DataBase\Contracts\Model;
-
-class KeyValue extends Model {
+class KeyValue extends AbstractWayfairModel
+{
 
   /**
    * @var string
@@ -35,7 +35,8 @@ class KeyValue extends Model {
   /**
    * @return string
    */
-  public function getTableName(): string {
+  public function getTableName(): string
+  {
     return 'Wayfair::KeyValue';
   }
