+++ src/Models/PendingOrders.php
@@ -1,13 +1,13 @@
 <?php
+
 /**
  * @copyright 2019 Wayfair LLC - All rights reserved
  */
 
-namespace Wayfair\Models;
+use Wayfair\Models\AbstractWayfairModel;
 
-use Plenty\Modules\Plugin\DataBase\Contracts\Model;
-
-class PendingOrders extends Model {
+class PendingOrders extends AbstractWayfairModel
+{
 
   /**
    * @var      int
@@ -32,7 +32,8 @@ class PendingOrders extends Model {
   /**
    * @return string
    */
-  public function getTableName(): string {
+  public function getTableName(): string
+  {
     return 'Wayfair::PendingOrders';
   }
